With new API

+++ PoGo.PokeMobBot.Logic/Tasks/CatchIncensePokemonsTask.cs
@@ -62,7 +62,7 @@ namespace PoGo.PokeMobBot.Logic.Tasks
 
                     var encounter =
                         await
-                            session.Client.Encounter.EncounterIncensePokemon((long) pokemon.EncounterId,
+                            session.Client.Encounter.EncounterIncensePokemon(pokemon.EncounterId,
                                 pokemon.SpawnPointId);
 
                     if (encounter.Result == IncenseEncounterResponse.Types.Result.IncenseEncounterSuccess)
